# Rewrite n_grams function to a 2x faster recursive version