Hi, 

This recursive n_gram implementation is 2x faster.
You can check the benchmark in my fork in the `bench_new_hamming` branch.
You can run the benchmark in that branch with the `mix run bench/bench_n_grams.exs` command.

:wave: 
```
Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 14 s

Benchmarking n_grams...
Benchmarking n_grams2...

Name               ips        average  deviation         median         99th %
n_grams2        2.91 K      343.69 μs    ±14.96%      328.34 μs      597.89 μs
n_grams         1.37 K      728.21 μs    ±15.53%      683.96 μs     1211.69 μs

Comparison: 
n_grams2        2.91 K
n_grams         1.37 K - 2.12x slower +384.52 μs
```